+++ src/pages/main-page/main-page.tsx
@@ -3,8 +3,7 @@ import {useAppDispatch, useAppSelector} from '../../hooks';
 import {AuthorizationStatus, RequestStatus} from '../../const';
 import {fetchOffersAction, fetchFavoritesAction} from '../../store/api-actions';
 import {getAuthorizationStatus} from '../../store/user-data/user-data.selectors';
-import {getOffersFetchingStatus} from '../../store/offers-data/offers-data.selectors';
-import {getOffers, getActiveCity} from '../../store/offers-data/offers-data.selectors';
+import {getOffers, getOffersFetchingStatus, getActiveCity} from '../../store/offers-data/offers-data.selectors';
 import {CitiesListMemo as CitiesList} from '../../components/cities-list/cities-list';
 import HeaderFull from '../../components/header/header-full';
 import MainEmpty from '../../components/main-empty/main-empty';
